took out the middle man in the company create components, made button conditionally render
