+++ public/index.html
@@ -9,10 +9,7 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <style>
-      @import url('https://fonts.googleapis.com/css2?family=Oswald:wght@200;400;600&display=swap');
-    </style>
-    <title>Company Review</title>
+  <title>Company Review</title>
   </head>
   <body>
     <div id="root"></div>
