+++ src/index.css
@@ -1,8 +1,6 @@
-
-
 body {
   margin: 0;
-  font-family: Oswald;
+
   /* -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale; */
 }
